Moved special calculations to the route-level
They can’t be on the model because it needs to occur in init but it also needs to peekAll of the records, which is an infinite loop. These are taken care of at the route-level in a poorly-named Mixin called “Downloadable”. This should be renamed to something clearer - its intent has changed completely.