Fixed data names for data to appear
Found that there's extra spaces in the database variable names for a few variables where data is not appearing in the application. Think I fixed this issue by adding a space to the variable name in the code. We will eventually have to fix this in the database and remove the spaces from the code.